pyproject.toml
[tool.poetry]
name = "airflow"
version = "0.1.0"
description = ""
authors = ["miintto <miintto.log@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
apache-airflow = "^2.2.4"
apache-airflow-providers-docker = "^2.5.2"
psycopg2 = "^2.9.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
